requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
easy_async_tg_notify==0.1
exceptiongroup==1.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-decouple==3.8
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3==2.3.0
